pre 0.6 updates
[Added] Dynamic Debris Epoch Event. Dynamically spawns vehicle and other
debris on the roadways.
Scans the roads on startup and uses a random seed generator to determine
the locations of debris.
Debris is spawned using createSimpleObject and is made to conform with
the upper most roadway surface.

[Added]  Epoch Events 3.0
- Supports events based on server side configs and pbos.
- New server side events can be added by just dropping .pbo into
@epochhive/addons folder.
- To make your own standalone events you can use Dynamic Debris Event as
a template.

[Removed] Vehicle simulation handler, so we can use A3 Dynamic
Simulation instead.
[Changed] Safe Vehicle and token checks simplified to only check the
vehicles a player is in.
[Changed] Safe Vehicle classes can now be configured via server settings
pbo CfgSecConf >> vehicles >> safeVehicles array.