CfgActions overriding - fire on/off req - a3 hooking ++
[FIXED] BurnBarrel and FirePlace02 UserActions requiring lighter_epoch
did not need lighter when performed.
[NEW] CfgActions event handler for "Action" ("PrevAction" and
"NextAction" - not implemented yet)
-Allows ability to override any of the CfgActions/UserActions in Arma 3
and other mods.
[ADDED] usedItemRepack function to remove 1 use from the lighter.
[ADDED] Require fire extinguisher to "Put Out Fire" on the Burn Barrel
[CHANGED] ItemFireExtinguisher now has count = 5 for 5 uses to put out a
small fire
[ADDED] ItemFireExtinguisher to CfgItemSort and CfgPricing
[ADDED] Arma Sling Loading requires ItemRope to sling and returns rope
upon release..
[ADDED] Checks for R3F can_lift and Advanced Slingload InitVar.
[ADDED] If above mod/script is present disable Arma SlingLoad
Hook/Unhook action and key press.
[CHANGED] While working with keyDown EH, debug logs show the DIKCodes
are being pushed as whole numbers and not the 0x01 as the escape keyDown
handler was setup to use and why my tests were failing using the 0x30
for the letter B(Hook/Unhook).
[UPGRADED] Changed over the 0x01 and 0x30 to the actionKeys name in
order to properly handle those with client using custom keys.
[TEASED] Have a commented line in the Light Fire case showing future
plans for a fire system(BURN BABY BURN!).

[NEW] EPOCH_usedItemRepack

Description:
Allows adjusting of "ammo" counts in a magazine.
Can be used locally and on the server for a remote unit.
```
[_player,_magazine,_change] call EPOCH_usedItemRepack;