Reload Loadout on Pause
Thx to Grahame for the inspiration to add it just before logout.
This should fix the login without clothings.
This will not fix not visible Clothings for others, because it only runs
just before logout, but Players can force it be pressing ESC and waiting
until this command is fired.