Changed case, changed PW for parse
I changed the case of the files in MP for linux dedicated servers and changed the reference to altis in server-example.cfg.
Added a bash script to tools to make files lower case.
Made changes to references to passwords to make them easer to replace with tools like sed and made each reference unique but consistent.